+++ src/app/loadout-builder/LoadoutBuilder.tsx
@@ -212,9 +212,19 @@ function LoadoutBuilder({
         lockedMods,
         lockedExoticHash,
         upgradeSpendTier,
+        lockItemEnergyType,
         searchFilter
       ),
-    [defs, characterItems, lockedMap, lockedMods, lockedExoticHash, upgradeSpendTier, searchFilter]
+    [
+      defs,
+      characterItems,
+      lockedMap,
+      lockedMods,
+      lockedExoticHash,
+      upgradeSpendTier,
+      lockItemEnergyType,
+      searchFilter,
+    ]
   );
 
   const { result, processing } = useProcess(